refactor(app): move the map page from locale] routing to a non locale one
