--- src/app/[locale]/page.tsx
@@ -1,3 +1,0 @@
-export default async function Home() {
-  return <h1>Landing page</h1>
-}
